backend/idca_func/requirements.txt
azure-ai-agents
azure-identity
azure-storage-blob
azure-data-tables
PyPDF2
httpx
python-dotenv
